Extract shopping list helpers
